When using non-catchall dynamic dns, machine A rec
Thanks to a quirk of how the dynamic dns plugin is implemented, a
domain/subdomain must be specified for the caddy instance, before
picking up automatic subdomains later.

As such, to satisfy this requirement without having multiple machines
"fighting" over a single A-name (which wouldn't really be a problem, but
seems silly), we can just set a dummy or sacrificial per-machine DNS
record.